updating user id from Integer to BigInteger

+++ Codigo/backend/src/models/db/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import BigInteger, Column, String
 
 from .base import Base
 
@@ -6,7 +6,7 @@ from .base import Base
 class UserModel(Base):
     __tablename__ = "users"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     email = Column(String, unique=True, index=True)
     hashed_password = Column(String)
     name = Column(String, nullable=False)
